+++ forge/src/main/java/com/jahirtrap/foodtxf/FoodtxfMod.java
@@ -4,6 +4,7 @@ import com.jahirtrap.configlib.TXFConfig;
 import com.jahirtrap.foodtxf.init.ModConfig;
 import com.jahirtrap.foodtxf.init.ModContent;
 import com.jahirtrap.foodtxf.init.ModTab;
+import com.jahirtrap.foodtxf.init.ModVillagerTrades;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
@@ -19,5 +20,6 @@ public class FoodtxfMod {
         TXFConfig.init(MODID, ModConfig.class);
         ModContent.init(bus);
         ModTab.init(bus);
+        ModVillagerTrades.init(bus);
     }
 }